1. 增加了 改贴删帖的验证，repair的验证（未测试）

+++ src/main/java/com/example/demo/controller/TieController.java
@@ -2,8 +2,12 @@ package com.example.demo.controller;
 
 import com.example.demo.dto.ResultDTO;
 import com.example.demo.model.Tie;
+import com.example.demo.service.NotificationService;
 import com.example.demo.service.TieService;
+import com.example.demo.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -13,10 +17,17 @@ public class TieController {
     @Autowired
     private TieService tieService;
 
+    @Autowired
+    private UserService userService;
+
+    @Autowired
+    private NotificationService notificationService;
+
 
     /* 发帖 */
     @PostMapping("/publish")
     public ResultDTO doPublish(@RequestBody Tie tie){
+
         ResultDTO result = tieService.publish(tie);
         return result;
     }
@@ -24,17 +35,37 @@ public class TieController {
     /* 删帖 */
     @DeleteMapping("/delete/{tieId}")
     public ResultDTO deDelete(@PathVariable("tieId") Integer tieId){
-        ResultDTO result = tieService.delete(tieId);
-        return result;
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();//当前登录的用户名
+        Integer userId = userService.getUserId(username);
+
+        if (tieService.findRepairByUserId(userId)){
+            ResultDTO result = tieService.delete(tieId);
+            return result;
+        }else{
+            return ResultDTO.errorOf(401,"你无权这么做");
+        }
+
+
     }
 
     /* 改贴 */
     @PutMapping("/update/{tieId}")
     public ResultDTO doUpdate(@RequestBody Tie tie,
                               @PathVariable("tieId") Integer tieId){
-        tie.setTieId(tieId);
-        ResultDTO update = tieService.update(tie);
-        return update;
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();//当前登录的用户名
+        Integer userId = userService.getUserId(username);
+
+        if (tieService.findRepairByUserId(userId)){
+            tie.setTieId(tieId);
+            ResultDTO update = tieService.update(tie);
+            return update;
+        }else{
+            return ResultDTO.errorOf(401,"你无权这么做");
+        }
     }
 
     /* 查询所有帖子 */
